refactor: split code into files
