+++ config.example.toml
@@ -30,3 +30,4 @@ use_grafana = true
 id = "PRECONF"
 type = "preconf"
 docker_image = "test_preconf"
+beacon_nodes = ["http://127.0.0.1:60158"]
